Created K-Means Prototype
kmeansprotoype.py contains three definitions and a few lines of test code. The brain of the operation is kmeans_cluster_prototype which takes the parameters k, points (array of points), and optionally the starting centroids (array of points). The function returns an array of all clusters and currently the program iterates for all values of k between 2 and 8.